+++ src/main/g8/app/controllers/IndexController.scala
@@ -13,7 +13,7 @@ class IndexController @Inject()(
                                  view: IndexView
                                ) extends FrontendBaseController with I18nSupport {
 
-  def onPageLoad: Action[AnyContent] = identify { implicit request =>
+  def onPageLoad(): Action[AnyContent] = identify { implicit request =>
     Ok(view())
   }
 }