+++ src/main/g8/app/controllers/UnauthorisedController.scala
@@ -11,7 +11,7 @@ class UnauthorisedController @Inject()(
                                         view: UnauthorisedView
                                       ) extends FrontendBaseController with I18nSupport {
 
-  def onPageLoad: Action[AnyContent] = Action { implicit request =>
+  def onPageLoad(): Action[AnyContent] = Action { implicit request =>
     Ok(view())
   }
 }