+++ src/main/g8/app/controllers/auth/SignedOutController.scala
@@ -11,7 +11,7 @@ class SignedOutController @Inject()(
                                      view: SignedOutView
                                    ) extends FrontendBaseController with I18nSupport {
 
-  def onPageLoad: Action[AnyContent] = Action { implicit request =>
+  def onPageLoad(): Action[AnyContent] = Action { implicit request =>
     Ok(view())
   }
 }