+++ src/main/g8/app/models/UserAnswers.scala
@@ -70,7 +70,7 @@ object UserAnswers {
       (__ \ "_id").write[String] and
       (__ \ "data").write[JsObject] and
       (__ \ "lastUpdated").write(MongoJavatimeFormats.instantFormat)
-    ) (unlift(UserAnswers.unapply))
+    ) (ua => (ua.id, ua.data, ua.lastUpdated))
   }
 
   implicit val format: OFormat[UserAnswers] = OFormat(reads, writes)