+++ src/main/g8/test/forms/behaviours/StringFieldBehaviours.scala
@@ -12,7 +12,7 @@ trait StringFieldBehaviours extends FieldBehaviours {
     s"not bind strings longer than \$maxLength characters" in {
 
       forAll(stringsLongerThan(maxLength) -> "longString") {
-        string =>
+        (string: String) =>
           val result = form.bind(Map(fieldName -> string)).apply(fieldName)
           result.errors must contain only lengthError
       }