+++ src/main/scaffolds/multipleQuestionsPage/app/models/$className$.scala
@@ -5,5 +5,6 @@ import play.api.libs.json._
 case class $className$ ($field1Name$: String, $field2Name$: String)
 
 object $className$ {
-  implicit val format = Json.format[$className$]
+  
+  implicit val format: OFormat[$className$] = Json.format
 }