+++ CHANGELOG.md
@@ -10,6 +10,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ### Changed
 * Update to SBT 1.9.9
+* Update dependencies
+* Consolidate syntax to make a future migration to scala 3 simpler
 
 ## v0.36.0 - 2024-05-31